Move file-stuff around a bit
There is no longer an `index` method in `FilesController`, instead
there is a browse method that serves json results. The old index
page has moved to the `new` method, which we call from the url
`/files/add` to be clearer on what this url does (that is, add
files from from a remote service).