The new method for files is now rendered with angular routes
All routes are specified in the main.coffee file. The base URL is necessary in the application.html.erb, otherwise angular would not recognise the entire url. So I added the @base_href variable to the new methdo so it wont be rendered with that base href all the time.

Moved the templates to other folders, so the new.html.erb file will be a bit easier to read.